+++ lib/src/eihpdef.h
@@ -1,6 +1,8 @@
 #ifndef EIHPDEF_H
 #define EIHPDEF_H
 
+#include <vms_types.h>
+
 #define EIHP$K_LENGTH 60
 #define EIHP$C_LENGTH 60
 #define EIHP$S_EIHPDEF 60
@@ -9,27 +11,27 @@ struct _eihp
 {
     struct
     {
-        unsigned int eihp$l_majorid;
-        unsigned int eihp$l_minorid;
+        UINT32 eihp$l_majorid;
+        UINT32 eihp$l_minorid;
     };
-    unsigned int eihp$l_eco1;
-    unsigned int eihp$l_eco2;
-    unsigned int eihp$l_eco3;
-    unsigned int eihp$l_eco4;
-    unsigned int eihp$l_patcomtxt;
-    unsigned int eihp$l_rw_patsiz;
+    UINT32 eihp$l_eco1;
+    UINT32 eihp$l_eco2;
+    UINT32 eihp$l_eco3;
+    UINT32 eihp$l_eco4;
+    UINT32 eihp$l_patcomtxt;
+    UINT32 eihp$l_rw_patsiz;
     union
     {
-        unsigned int eihp$l_rw_patadr;
-        unsigned long long eihp$q_rw_patadr;
+        UINT32 eihp$l_rw_patadr;
+        UINT64 eihp$q_rw_patadr;
     };
-    unsigned int eihp$l_ro_patsiz;
+    UINT32 eihp$l_ro_patsiz;
     union
     {
-        unsigned int eihp$l_ro_patadr;
-        unsigned long long eihp$q_ro_patadr;
+        UINT32 eihp$l_ro_patadr;
+        UINT64 eihp$q_ro_patadr;
     };
-    unsigned long long eihp$q_patdate;
+    UINT64 eihp$q_patdate;
 };
 
 #endif
